fix: fix api route
